refactors, move electrs / monerod / asb module into meta containers module
