+++ swap-orchestrator/src/lib.rs
@@ -1,6 +1,3 @@
-pub mod asb;
-pub mod bitcoind;
 pub mod compose;
-pub mod electrs;
+pub mod containers;
 pub mod images;
-pub mod monerod;
